Simplified further by merging Instance as an inner class of Logistic and moving the static method from DataSet to Logistic.
